+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * `route.print` not handle keyword matching routes correctly #1136
 * Allow developer to disable JSON body patch #1134
 * API Doc - treat session variable as a special param #1133
+* API doc - catch field javadoc comment and add it to API doc #1132
 * Enhance AdaptiveBean - better support to inheritance #1131
 * Allow developer to specify property naming convention for JSON output #1130
 * Upload file issue #1128