+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # ActFramework Change Log
 
-**1.8.22**
+**1.8.22** 20/May/2019
 * `route.print` not handle keyword matching routes correctly #1136
 * Allow developer to disable JSON body patch #1134
 * API Doc - treat session variable as a special param #1133